move some tests around, fix imports and rename old tests
