+++ tests/test_readers.py
@@ -1,3 +1,4 @@
+from os.path import join
 from pathlib import Path
 
 import numpy as np
@@ -8,6 +9,7 @@ from hydromt_sfincs.readers import (
     read_binary_map_index,
     read_config,
     read_geoms,
+    read_xy,
 )
 
 from .conftest import TESTMODELDIR
@@ -70,3 +72,12 @@ def test_read_geoms():
     assert 322500 < np.mean(g[0]["x"]) < 322750
     assert g[0]["elevation"] == [3.5] + [3.0] * 9
     assert g[0]["par1"] == [0.6] * 10
+
+
+def test_read_xy_index_is_zero_based(tmp_dir):
+    """read_xy returns a canonical 0-based index (bnd/src points)."""
+    fn = join(tmp_dir, "sfincs.bnd")
+    with open(fn, "w") as f:
+        f.write("0.0 0.0\n10.0 0.0\n10.0 10.0\n")
+    gdf = read_xy(fn, crs=32633)
+    assert list(gdf.index) == [0, 1, 2]
