+++ tests/test_utils.py
@@ -1,10 +1,9 @@
 """Test sfincs utils"""
 
-import pytest
-import numpy as np
-from shapely.geometry import MultiLineString, Point
 import geopandas as gpd
-import copy
+import numpy as np
+import pytest
+from shapely.geometry import MultiLineString, Point
 
 from hydromt_sfincs import utils
 
@@ -77,31 +76,3 @@ def test_make_regular_grid(rotation, uv_points):
     # assertions
     np.testing.assert_allclose(da_transform, transform, atol=1e-8)
     assert da.shape == (height, width)
-
-
-def test_read_xy_index_is_zero_based(tmp_dir):
-    """read_xy returns a canonical 0-based index (bnd/src points)."""
-    fn = join(tmp_dir, "sfincs.bnd")
-    with open(fn, "w") as f:
-        f.write("0.0 0.0\n10.0 0.0\n10.0 10.0\n")
-    gdf = utils.read_xy(fn, crs=32633)
-    assert list(gdf.index) == [0, 1, 2]
-
-
-def test_write_xyn_drops_z_of_3d_points(tmp_dir):
-    """write_xyn must tolerate 3D (X, Y, Z) geometries from GIS by dropping Z."""
-    from shapely.geometry import Point
-
-    gdf = gpd.GeoDataFrame(
-        {"name": ["a", "b"]},
-        geometry=[Point(1.0, 2.0, 99.0), Point(3.0, 4.0, 88.0)],
-        crs=32633,
-    )
-    fn = join(tmp_dir, "sfincs.obs")
-    utils.write_xyn(fn, gdf)  # must not raise on the Z coordinate
-    gdf_rt = utils.read_xyn(fn)
-    assert len(gdf_rt) == 2
-    np.testing.assert_allclose(
-        [gdf_rt.geometry.x.tolist(), gdf_rt.geometry.y.tolist()],
-        [[1.0, 3.0], [2.0, 4.0]],
-    )
